verify test works
